Split views and helpers to more manageable file structure
