--- danceschool/financial/helpers.py
@@ -1,1400 +1,0 @@
-from django.conf import settings
-from django.db.models import (
-    Sum, Count, Q, ExpressionWrapper, F, DurationField, Case, When, Min
-)
-from django.db.models.functions import TruncDate, TruncMonth, Coalesce
-from django.http import HttpResponse
-from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.contrib.contenttypes.models import ContentType
-
-from datetime import datetime, timedelta
-from dateutil.relativedelta import relativedelta
-import unicodecsv as csv
-from calendar import month_name
-import pytz
-
-from danceschool.core.constants import getConstant
-from danceschool.core.models import (
-    Registration, Event, EventOccurrence, EventStaffMember, InvoiceItem, Room
-)
-from danceschool.core.utils.timezone import ensure_timezone, ensure_localtime
-
-from .constants import EXPENSE_BASES
-from .models import (
-    ExpenseItem, RevenueItem, RepeatedExpenseRule, RoomRentalInfo,
-    TransactionParty, ExpensePurpose
-)
-
-
-def getExpenseItemsCSV(queryset, scope='instructor'):
-
-    response = HttpResponse(content_type='text/csv')
-    if scope == 'instructor':
-        response['Content-Disposition'] = 'attachment; filename="paymentHistory.csv"'
-    else:
-        response['Content-Disposition'] = 'attachment; filename="expenseHistory.csv"'
-
-    writer = csv.writer(response, csv.excel)
-    # BOM (optional...Excel needs it to open UTF-8 file properly)
-    response.write(u'\ufeff'.encode('utf8'))
-
-    header_list = [
-        _('Description'),
-        _('Expense Category'),
-        _('Hours'),
-        _('Wage Rate'),
-        _('Total Payment'),
-        _('Is Reimbursement'),
-        _('Submission Date'),
-        _('Event'),
-        _('Approved'),
-        _('Approval Date'),
-        _('Paid'),
-        _('Payment Date'),
-        _('Payment Method'),
-        _('Accrual Date'),
-    ]
-
-    if scope != 'instructor':
-        header_list += [_('Pay To')]
-
-    writer.writerow(header_list)
-
-    for x in queryset:
-        this_row_data = [
-            x.description,
-            x.category.name,
-            x.hours,
-            x.wageRate,
-            x.total,
-            x.reimbursement,
-            x.submissionDate,
-            x.event,
-            x.approved,
-            x.approvalDate,
-            x.paid,
-            x.paymentDate,
-            x.paymentMethod,
-            x.accrualDate,
-        ]
-
-        if scope != 'instructor':
-            this_row_data.append(x.payTo)
-
-        writer.writerow(this_row_data)
-    return response
-
-
-def getRevenueItemsCSV(queryset):
-    response = HttpResponse(content_type='text/csv')
-    response['Content-Disposition'] = 'attachment; filename="revenueHistory.csv"'
-
-    writer = csv.writer(response, csv.excel)
-    # BOM (optional...Excel needs it to open UTF-8 file properly)
-    response.write(u'\ufeff'.encode('utf8'))
-
-    header_list = [
-        _('Description'),
-        _('Revenue Category'),
-        _('Gross Total (Pre-Discounts & Vouchers)'),
-        _('Net Total'),
-        _('Received From'),
-        _('Invoice ID'),
-        _('Event'),
-        _('Submission Date'),
-        _('Received'),
-        _('Received Date'),
-        _('Payment Method'),
-        _('Accrual Date'),
-    ]
-    writer.writerow(header_list)
-
-    for x in queryset:
-        this_row_data = [
-            x.description,
-            x.category.name,
-            x.grossTotal,
-            x.total,
-            getattr(x.receivedFrom, 'name', None),
-            getattr(getattr(x.invoiceItem, 'invoice', None), 'id', None),
-            x.event,
-            x.submissionDate,
-            x.received,
-            x.receivedDate,
-            x.paymentMethod,
-            x.accrualDate,
-        ]
-
-        writer.writerow(this_row_data)
-    return response
-
-def getExpenseCategoryForStaffer(staffer=None, staffCategory=None):
-    '''
-    For standard categories of staff, map the EventStaffCategory to
-    an ExpenseCategory using the stored constants.  Otherwise, the
-    ExpenseCategory is a generic one.
-    '''
-    if staffer:
-        staffCategory=staffer.category
-
-    if staffCategory == getConstant('general__eventStaffCategoryAssistant'):
-        return getConstant('financial__assistantClassInstructionExpenseCat')
-    elif staffCategory in [
-            getConstant('general__eventStaffCategoryInstructor'),
-            getConstant('general__eventStaffCategorySubstitute')
-    ]:
-        return getConstant('financial__classInstructionExpenseCat')
-    else:
-        # This is the generic category for all Event staff, but it may be overridden below
-        return getConstant('financial__otherStaffExpenseCat')
-
-
-def getBestRuleForStaffer(staffer=None, staffCategory=None, staffMember=None):
-    '''
-    When generating expense items...
-    '''
-    pass
-
-
-def addEventTimeFilters(
-    event_timefilters=None, rule=None, datetimeTuple=None, event=None
-):
-    '''
-    This function provides the filters needed to identify events for which an
-    expense item may need to be generated.
-    '''
-
-    if not event_timefilters:
-        event_timefilters = Q()
-
-    c = getConstant('financial__autoGenerateEventExpensesWindow') or 30
-    event_timefilters = event_timefilters & Q(
-        event__endTime__gte=(timezone.now() - timedelta(days=c))
-    )
-
-    if datetimeTuple and len(datetimeTuple) == 2:
-        timelist = list(datetimeTuple)
-        timelist.sort()
-        event_timefilters = event_timefilters & (
-            Q(event__startTime__gte=timelist[0]) & Q(event__startTime__lte=timelist[1])
-        )
-
-    if event:
-        event_timefilters = event_timefilters & Q(event__id=event.id)
-
-    if rule:
-        if rule.advanceDays is not None:
-            if rule.advanceDaysReference == RepeatedExpenseRule.MilestoneChoices.end:
-                event_timefilters = event_timefilters & \
-                    Q(event__endTime__lte=timezone.now() + timedelta(days=rule.advanceDays))
-            elif rule.advanceDaysReference == RepeatedExpenseRule.MilestoneChoices.start:
-                event_timefilters = event_timefilters & \
-                    Q(event__startTime__lte=timezone.now() + timedelta(days=rule.advanceDays))
-        if rule.priorDays is not None:
-            if rule.priorDaysReference == RepeatedExpenseRule.MilestoneChoices.end:
-                event_timefilters = event_timefilters & \
-                    Q(event__endTime__gte=timezone.now() - timedelta(days=rule.priorDays))
-            elif rule.priorDaysReference == RepeatedExpenseRule.MilestoneChoices.start:
-                event_timefilters = event_timefilters & \
-                    Q(event__startTime__gte=timezone.now() - timedelta(days=rule.priorDays))
-        if rule.startDate:
-            event_timefilters = event_timefilters & Q(
-                event__startTime__gte=timezone.now().replace(
-                    year=rule.startDate.year, month=rule.startDate.month, day=rule.startDate.day,
-                    hour=0, minute=0, second=0, microsecond=0,
-                )
-            )
-        if rule.endDate:
-            event_timefilters = event_timefilters & Q(
-                event__startTime__lte=timezone.now().replace(
-                    year=rule.endDate.year, month=rule.endDate.month, day=rule.endDate.day,
-                    hour=0, minute=0, second=0, microsecond=0,
-                )
-            )
-
-    return event_timefilters
-
-def createExpenseItemsForVenueRental(request=None, datetimeTuple=None, rule=None, event=None):
-    '''
-    For each Location or Room-related Repeated Expense Rule, look for Events
-    in the designated time window that do not already have expenses associated
-    with them.  For hourly rental expenses, then generate new expenses that are
-    associated with this rule.  For non-hourly expenses, generate new expenses
-    based on the non-overlapping intervals of days, weeks or months for which
-    there is not already an ExpenseItem associated with the rule in question.
-    '''
-
-    # These are used repeatedly, so they are put at the top
-    submissionUser = getattr(request, 'user', None)
-    rental_category = getConstant('financial__venueRentalExpenseCat')
-
-    # Return the number of new expense items created
-    generate_count = 0
-
-    # First, construct the set of rules that need to be checked for affiliated events
-    rule_filters = Q(disabled=False) & Q(rentalRate__gt=0) & \
-        (Q(locationrentalinfo__isnull=False) | Q(roomrentalinfo__isnull=False))
-    if rule:
-        rule_filters = rule_filters & Q(id=rule.id)
-    rulesToCheck = RepeatedExpenseRule.objects.filter(rule_filters).distinct()
-
-    # Now, we loop through the set of rules that need to be applied, then loop through the
-    # Events in the window in question that occurred at the location indicated by the rule.
-    for rule in rulesToCheck:
-
-        venue = (
-            getattr(rule, 'location', None) if
-            isinstance(rule, RoomRentalInfo) else
-            getattr(rule, 'location', None)
-        )
-        loc = getattr(venue, 'location') if isinstance(venue, Room) else venue
-        event_locfilter = Q(event__room=venue) if isinstance(venue, Room) else Q(event__location=venue)
-
-        # Find or create the TransactionParty associated with the location.
-        loc_party = TransactionParty.objects.get_or_create(
-            location=loc, defaults={'name': loc.name}
-        )[0]
-
-        event_timefilters = addEventTimeFilters(
-            rule=rule, datetimeTuple=datetimeTuple, event=event
-        )
-
-        # For construction of expense descriptions
-        replacements = {
-            'type': _('Event/Series venue rental'),
-            'of': _('of'),
-            'location': venue.name,
-            'for': _('for'),
-        }
-
-        # Loop through EventOccurrences for which there are not already directly
-        # allocated expenses under this rule, and create new ExpenseItems for
-        # them depending on whether the rule requires hourly expenses or
-        # non-hourly ones to be generated.
-        occurrences = EventOccurrence.objects.filter(
-            event_locfilter & event_timefilters
-        ).exclude(Q(related_expenses__item__expenseRule=rule)).select_related(
-            'event', 'event__location', 'event__room'
-        ).distinct()
-
-        if rule.applyRateRule == rule.RateRuleChoices.hourly:
-            for this_occ in occurrences:
-                # Hourly expenses are always generated without checking for
-                # overlapping windows, because the periods over which hourly expenses
-                # are defined are disjoint.  However, hourly expenses are allocated
-                # directly to events, so we just need to create expenses for any events
-                # that do not already have an Expense Item generate under this rule.
-                replacements['name'] = this_occ.event.name
-                replacements['dates'] = this_occ.localStartTime.strftime('%Y-%m-%d')
-                if (
-                        this_occ.localStartTime.strftime('%Y-%m-%d') !=
-                        this_occ.localEndTime.strftime('%Y-%m-%d')
-                ):
-                    replacements['dates'] += ' %s %s' % (
-                        _('to'), this_occ.localEndTime.strftime('%Y-%m-%d')
-                    )
-
-                new_item = ExpenseItem.objects.create(
-                    event=this_occ.event,
-                    category=rental_category,
-                    payTo=loc_party,
-                    expenseRule=rule,
-                    description=(
-                        '%(type)s %(of)s %(location)s %(for)s: %(name)s, %(dates)s' %
-                        replacements
-                    ),
-                    submissionUser=submissionUser,
-                    total=this_occ.duration * rule.rentalRate,
-                    accrualDate=this_occ.startTime,
-                )
-
-                # Record that this occurrence was the purpose of the
-                # newly-generated expense item.
-                ExpensePurpose.objects.create(
-                    item=new_item,
-                    purpose=this_occ
-                )
-
-                generate_count += 1
-        else:
-            # Non-hourly expenses are generated by constructing the time
-            # intervals in which the occurrence occurs, and removing from that
-            # interval any intervals in which an expense has already been
-            # generated under this rule (so, for example, monthly rentals will
-            # now show up multiple times). So, we just need to construct the set
-            # of intervals for which to construct expenses
-            intervals = [(x.localStartTime, x.localEndTime) for x in occurrences]
-            remaining_intervals = rule.getWindowsAndTotals(
-                intervals, remove_existing_overlaps=True
-            )
-
-            for startTime, endTime, total, description in remaining_intervals:
-                replacements['when'] = description
-
-                new_item, created = ExpenseItem.objects.get_or_create(
-                    category=rental_category,
-                    payTo=loc_party,
-                    expenseRule=rule,
-                    periodStart=startTime,
-                    periodEnd=endTime,
-                    defaults={
-                        'description': '%(type)s %(of)s %(location)s %(for)s %(when)s' % replacements,
-                        'submissionUser': submissionUser,
-                        'total': total,
-                        'accrualDate': startTime,   
-                    }
-                )
-
-                # Record that this occurrence was the purpose of the
-                # newly-generated expense item.
-                ExpensePurpose.objects.create(
-                    item=new_item,
-                    purpose=this_occ
-                )
-
-                if created:
-                    generate_count += 1
-    rulesToCheck.update(lastRun=timezone.now())
-    return generate_count
-
-
-def createExpenseItemsForEvents(request=None, datetimeTuple=None, rule=None, event=None):
-    '''
-    For each StaffMember-related Repeated Expense Rule, look for EventStaffMember
-    instances in the designated time window that do not already have expenses associated
-    with them.  For hourly rental expenses, then generate new expenses that are
-    associated with this rule.  For non-hourly expenses, generate new expenses
-    based on the non-overlapping intervals of days, weeks or months for which
-    there is not already an ExpenseItem associated with the rule in question.
-    '''
-
-    # This is used repeatedly, so it is put at the top
-    submissionUser = getattr(request, 'user', None)
-
-    # Return the number of new expense items created
-    generate_count = 0
-
-    # First, construct the set of rules that need to be checked for affiliated events
-    rule_filters = Q(disabled=False) & Q(rentalRate__gt=0) & \
-        Q(Q(staffmemberwageinfo__isnull=False) | Q(staffdefaultwage__isnull=False))
-    if rule:
-        rule_filters = rule_filters & Q(id=rule.id)
-    rulesToCheck = RepeatedExpenseRule.objects.filter(
-        rule_filters).distinct().order_by(
-            '-staffmemberwageinfo__category', '-staffdefaultwage__category'
-        )
-
-    # Now, we loop through the set of rules that need to be applied, then loop
-    # through the Events in the window in question that involved the staff
-    # member indicated by the rule.
-    for rule in rulesToCheck:
-        staffMember = getattr(rule, 'staffMember', None)
-        staffCategory = getattr(rule, 'category', None)
-
-        # No need to continue if expenses are not to be generated
-        if (
-                (not staffMember and not staffCategory) or
-                (
-                    not staffMember and not
-                    getConstant('financial__autoGenerateFromStaffCategoryDefaults')
-                )
-        ):
-            continue
-
-        # For construction of expense descriptions
-        replacements = {
-            'type': _('Staff'),
-            'to': _('payment to'),
-            'for': _('for'),
-        }
-
-        if staffCategory:
-            if staffMember:
-                # This staff member in this category
-                eventstaff_filter = Q(staffMember=staffMember) & Q(category=staffCategory)
-            elif getConstant('financial__autoGenerateFromStaffCategoryDefaults'):
-                # Any staff member who does not already have a rule specified this category
-                eventstaff_filter = (
-                    Q(category=staffCategory) &
-                    ~Q(staffMember__expenserules__category=staffCategory)
-                )
-            replacements['type'] = staffCategory.name
-
-        else:
-            # We don't want to generate duplicate expenses when there is both a category-limited
-            # rule and a non-limited rule for the same person, so we have to construct the list
-            # of categories that are to be excluded if no category is specified by this rule.
-            coveredCategories = list(staffMember.expenserules.filter(
-                category__isnull=False).values_list('category__id', flat=True))
-            eventstaff_filter = Q(staffMember=staffMember) & ~Q(category__id__in=coveredCategories)
-
-        event_timefilters = addEventTimeFilters(
-            rule=rule, datetimeTuple=datetimeTuple, event=event
-        )
-
-        # Loop through EventStaffMembers for which there are not already
-        # directly allocated expenses under this rule, and create new
-        # ExpenseItems for them depending on whether the rule requires hourly
-        # expenses or non-hourly ones to be generated. Note: we don't prefetch
-        # the event on purpose, because this ends up being a non-polymorphic
-        # linkage, which means that subclass properties (like name) are not
-        # correctly handled.
-        staffers = EventStaffMember.objects.filter(
-            eventstaff_filter & event_timefilters
-        ).exclude(Q(related_expenses__item__expenseRule=rule)).select_related(
-            'staffMember',
-        ).prefetch_related('occurrences').distinct()
-
-        if rule.applyRateRule == rule.RateRuleChoices.hourly:
-            if getConstant('financial__autoGenerateExpensesEventStaff') == 'per_occurrence':
-                # Per-occurrence mode: create one ExpenseItem per staffer per
-                # occurrence, with hours proportional to occurrence duration.
-                # Re-query without the whole-staffer exclusion so we can check
-                # at the individual occurrence level inside the loop.
-                staffers_per_occ = EventStaffMember.objects.filter(
-                    eventstaff_filter & event_timefilters
-                ).select_related('staffMember').prefetch_related('occurrences').distinct()
-
-                for staffer in staffers_per_occ:
-                    allocation = staffer.allocationByOccurrence
-                    relevant_occs = (
-                        staffer.occurrences.filter(cancelled=False) or
-                        staffer.event.eventoccurrence_set.filter(cancelled=False)
-                    )
-
-                    staffer_party = TransactionParty.objects.get_or_create(
-                        staffMember=staffer.staffMember,
-                        defaults={
-                            'name': staffer.staffMember.fullName,
-                            'user': getattr(staffer.staffMember, 'userAccount', None)
-                        }
-                    )[0]
-
-                    for occ in relevant_occs:
-                        # Skip if an expense already exists for this
-                        # staffer + occurrence + rule combination.
-                        if staffer.related_expenses.filter(
-                            item__expenseRule=rule,
-                            occurrence=occ,
-                        ).exists():
-                            continue
-
-                        occ_hours = (
-                            allocation.get((occ.id, staffer.event.id), {}).get('duration', 0) / 3600
-                        )
-                        if not occ_hours:
-                            continue
-
-                        replacements['event'] = staffer.event.name
-                        replacements['name'] = staffer.staffMember.fullName
-                        replacements['dates'] = occ.localStartTime.strftime('%Y-%m-%d')
-
-                        new_item = ExpenseItem.objects.create(
-                            event=staffer.event,
-                            category=getExpenseCategoryForStaffer(staffer),
-                            expenseRule=rule,
-                            description='%(type)s %(to)s %(name)s %(for)s: %(event)s, %(dates)s' % replacements,
-                            submissionUser=submissionUser,
-                            hours=occ_hours,
-                            wageRate=rule.rentalRate,
-                            total=occ_hours * rule.rentalRate,
-                            accrualDate=occ.startTime,
-                            payTo=staffer_party,
-                        )
-
-                        ExpensePurpose.objects.create(
-                            item=new_item,
-                            purpose=staffer,
-                            occurrence=occ,
-                        )
-
-                        generate_count += 1
-            else:
-                # Per-event mode (default): one ExpenseItem per EventStaffMember.
-                for staffer in staffers:
-                    # Hourly expenses are always generated without checking for
-                    # overlapping windows, because the periods over which hourly
-                    # expenses are defined are disjoint.  However, hourly expenses
-                    # are allocated directly to events, so we just need to create
-                    # expenses for any events that do not already have an Expense
-                    # Item generated under this rule.
-                    replacements['event'] = staffer.event.name
-                    replacements['name'] = staffer.staffMember.fullName
-                    replacements['dates'] = staffer.event.localStartTime.strftime('%Y-%m-%d')
-                    if (
-                            staffer.event.localStartTime.strftime('%Y-%m-%d') !=
-                            staffer.event.localEndTime.strftime('%Y-%m-%d')
-                    ):
-                        replacements['dates'] += ' %s %s' % (
-                            _('to'), staffer.event.localEndTime.strftime('%Y-%m-%d')
-                        )
-
-                    # Find or create the TransactionParty associated with the staff member.
-                    staffer_party = TransactionParty.objects.get_or_create(
-                        staffMember=staffer.staffMember,
-                        defaults={
-                            'name': staffer.staffMember.fullName,
-                            'user': getattr(staffer.staffMember, 'userAccount', None)
-                        }
-                    )[0]
-
-                    params = {
-                        'event': staffer.event,
-                        'category': getExpenseCategoryForStaffer(staffer),
-                        'expenseRule': rule,
-                        'description': '%(type)s %(to)s %(name)s %(for)s: %(event)s, %(dates)s' % replacements,
-                        'submissionUser': submissionUser,
-                        'hours': staffer.netHours,
-                        'wageRate': rule.rentalRate,
-                        'total': staffer.netHours * rule.rentalRate,
-                        'accrualDate': staffer.event.startTime,
-                        'payTo': staffer_party,
-                    }
-
-                    new_item = ExpenseItem.objects.create(**params)
-
-                    # Record that this staffing was the purpose of the
-                    # newly-generated expense item.
-                    ExpensePurpose.objects.create(
-                        item=new_item,
-                        purpose=staffer
-                    )
-
-                    generate_count += 1
-        else:
-            # Non-hourly expenses are generated by constructing the time
-            # intervals in which the occurrence occurs, and removing from that
-            # interval any intervals in which an expense has already been
-            # generated under this rule (so, for example, monthly rentals will
-            # not show up multiple times). So, we just need to construct the set
-            # of intervals for which to construct expenses. This is done
-            # separately for each EventStaffMember instance so that we can keep
-            # track of the full set of purposes of each expense.
-            for staffer in staffers:
-
-                expense_category=getExpenseCategoryForStaffer(staffer)
-
-                # Find or create the TransactionParty associated with the staff member.
-                staffer_party = TransactionParty.objects.get_or_create(
-                    staffMember=staffer.staffMember,
-                    defaults={
-                        'name': staffer.staffMember.fullName,
-                        'user': getattr(staffer.staffMember, 'userAccount', None)
-                    }
-                )[0]                
-
-                # Intervals are defined by the occurrences for which the
-                # individual staffed, or by the full set of event occurrences
-                # if no staffing-specific occurrences are specified.
-                intervals = [
-                    (x.localStartTime, x.localEndTime) for x in
-                    (
-                        staffer.occurrences.all() or
-                        staffer.event.eventoccurrence_set.all()
-                    )
-                ]
-                remaining_intervals = rule.getWindowsAndTotals(
-                    intervals, remove_existing_overlaps=True,
-                    category=expense_category, payTo=staffer_party
-                )
-
-                for startTime, endTime, total, description in remaining_intervals:
-                    replacements['when'] = description
-                    replacements['name'] = staffer.staffMember.fullName
-
-                    new_item, created = ExpenseItem.objects.get_or_create(
-                        category=expense_category,
-                        payTo=staffer_party,
-                        expenseRule=rule,
-                        periodStart=startTime,
-                        periodEnd=endTime,
-                        defaults={
-                            'description': '%(type)s %(to)s %(name)s %(for)s %(when)s' % replacements,
-                            'submissionUser': submissionUser,
-                            'total': total,
-                            'accrualDate': startTime,
-                        }
-                    )
-
-                    # Record that this staffing was the purpose of the
-                    # newly-generated expense item.
-                    ExpensePurpose.objects.create(
-                        item=new_item,
-                        purpose=staffer
-                    )
-
-                    if created:
-                        generate_count += 1
-    rulesToCheck.update(lastRun=timezone.now())
-    return generate_count
-
-
-def createGenericExpenseItems(request=None, datetimeTuple=None, rule=None):
-    '''
-    Generic repeated expenses are created by just entering an
-    expense at each exact point specified by the rule, without
-    regard for whether events are scheduled in the specified
-    window,
-    '''
-
-    # These are used repeatedly, so they are put at the top
-    submissionUser = getattr(request, 'user', None)
-
-    # Return the number of new expense items created
-    generate_count = 0
-
-    # First, construct the set of rules that need to be checked for affiliated events
-    rule_filters = Q(disabled=False) & Q(rentalRate__gt=0) & \
-        Q(genericrepeatedexpense__isnull=False)
-    if rule:
-        rule_filters = rule_filters & Q(id=rule.id)
-    rulesToCheck = RepeatedExpenseRule.objects.filter(rule_filters).distinct()
-
-    # These are the filters place on Events that overlap the window in which
-    # expenses are being generated.
-    if datetimeTuple and len(datetimeTuple) == 2:
-        timelist = list(datetimeTuple)
-        timelist.sort()
-    else:
-        timelist = None
-
-    # Now, we loop through the set of rules that need to be applied, check for an
-    # existing expense item at each point specified by the rule, and create a new
-    # expense if one does not exist.
-    for rule in rulesToCheck:
-
-        limits = timelist or [ensure_timezone(datetime.min), ensure_timezone(datetime.max)]
-
-        if rule.advanceDays:
-            limits[1] = min(limits[1], timezone.now() + timedelta(days=rule.advanceDays))
-        if rule.priorDays:
-            limits[0] = max(limits[0], timezone.now() - timedelta(days=rule.priorDays))
-
-        if rule.startDate:
-            limits[0] = max(
-                limits[0],
-                timezone.now().replace(
-                    year=rule.startDate.year, month=rule.startDate.month, day=rule.startDate.day,
-                    hour=0, minute=0, second=0, microsecond=0,
-                )
-            )
-        if rule.endDate:
-            limits[1] = min(
-                limits[1],
-                timezone.now().replace(
-                    year=rule.endDate.year, month=rule.endDate.month, day=rule.endDate.day,
-                    hour=0, minute=0, second=0, microsecond=0,
-                )
-            )
-
-        # Find the first start time greater than the lower bound time.
-        if rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.hourly:
-            this_time = limits[0].replace(minute=0, second=0, microsecond=0)
-            if this_time < limits[0]:
-                this_time += timedelta(hours=1)
-        elif rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.daily:
-            this_time = limits[0].replace(hour=rule.dayStarts, minute=0, second=0, microsecond=0)
-            if this_time < limits[0]:
-                this_time += timedelta(days=1)
-        elif rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.weekly:
-            offset = limits[0].weekday() - rule.weekStarts
-            this_time = limits[0].replace(
-                day=limits[0].day - offset, hour=rule.dayStarts, minute=0, second=0, microsecond=0
-            )
-            if this_time < limits[0]:
-                this_time += timedelta(days=7)
-        else:
-            this_time = limits[0].replace(
-                day=rule.monthStarts, hour=rule.dayStarts, minute=0, second=0, microsecond=0
-            )
-            if this_time < limits[0]:
-                this_time += relativedelta(months=1)
-
-        while this_time <= limits[1]:
-            defaults_dict = {
-                'category': rule.category,
-                'description': rule.name,
-                'submissionUser': submissionUser,
-                'total': rule.rentalRate,
-                'accrualDate': this_time,
-                'payTo': rule.payTo,
-                'approved': str(_('Approved')) if rule.markApproved else None,
-                'paid': rule.markPaid,
-            }
-            item, created = ExpenseItem.objects.get_or_create(
-                expenseRule=rule,
-                periodStart=this_time,
-                periodEnd=this_time,
-                defaults=defaults_dict
-            )
-            if created:
-                generate_count += 1
-            if rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.hourly:
-                this_time += timedelta(hours=1)
-            elif rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.daily:
-                this_time += timedelta(days=1)
-            elif rule.applyRateRule == RepeatedExpenseRule.RateRuleChoices.weekly:
-                this_time += timedelta(days=7)
-            else:
-                this_time += relativedelta(months=1)
-    rulesToCheck.update(lastRun=timezone.now())
-    return generate_count
-
-
-def createRevenueItemsForRegistrations(request=None, datetimeTuple=None):
-
-    if hasattr(request, 'user'):
-        submissionUser = request.user
-    else:
-        submissionUser = None
-
-    this_category = getConstant('financial__registrationsRevenueCat')
-
-    filters_events = {
-        'revenueitem__isnull': True,
-        'eventRegistration__isnull': False,
-        'eventRegistration__registration__final': True,
-    }
-
-    if datetimeTuple:
-        timelist = list(datetimeTuple)
-        timelist.sort()
-
-        filters_events[
-            'eventRegistration__event__eventoccurrence__startTime__gte'
-        ] = timelist[0]
-        filters_events[
-            'eventRegistration__event__eventoccurrence__startTime__lte'
-        ] = timelist[1]
-    else:
-        c = getConstant('financial__autoGenerateRevenueRegistrationsWindow') or 0
-        if c > 0:
-            filters_events[
-                'eventRegistration__event__eventoccurrence__startTime__gte'
-            ] = timezone.now() - relativedelta(months=c)
-
-    for item in InvoiceItem.objects.filter(**filters_events).distinct():
-        if item.invoice.registration.paidOnline:
-            received = True
-        else:
-            received = False
-
-        revenue_description = _('Event Registration ') + \
-            str(item.eventRegistration.id) + ': ' + \
-            item.invoice.registration.fullName
-        RevenueItem.objects.create(
-            invoiceItem=item,
-            category=this_category,
-            description=revenue_description,
-            submissionUser=submissionUser,
-            grossTotal=item.grossTotal,
-            total=item.total,
-            received=received,
-            receivedDate=item.invoice.modifiedDate
-        )
-
-
-def prepareFinancialStatement(year=None):
-
-    if year:
-        filter_year = year
-    else:
-        filter_year = timezone.now().year
-
-    expenses_ytd = list(
-        ExpenseItem.objects.filter(accrualDate__year=filter_year).aggregate(Sum('total')).values()
-    )[0]
-    revenues_ytd = list(
-        RevenueItem.objects.filter(accrualDate__year=filter_year).aggregate(Sum('total')).values()
-    )[0]
-    expenses_awaiting_approval = list(
-        ExpenseItem.objects.filter(
-            Q(paid=False) & (Q(approved__isnull=True) | Q(approved__exact=''))
-        ).aggregate(Sum('total')).values()
-    )[0]
-    expenses_awaiting_payment = list(
-        ExpenseItem.objects.filter(
-            Q(paid=False) & ~(Q(approved__isnull=True) | Q(approved__exact=''))
-        ).aggregate(Sum('total')).values()
-    )[0]
-    expenses_paid_notapproved = list(
-        ExpenseItem.objects.filter(
-            Q(paid=True) & (Q(approved__isnull=True) | Q(approved__exact=''))
-        ).aggregate(Sum('total')).values()
-    )[0]
-
-    return {
-        'expenses_ytd': expenses_ytd,
-        'revenues_ytd': revenues_ytd,
-        'expenses_awaiting_approval': expenses_awaiting_approval,
-        'expenses_awaiting_payment': expenses_awaiting_payment,
-        'expenses_paid_notapproved': expenses_paid_notapproved,
-    }
-
-
-def prepareStatementByPeriod(**kwargs):
-    basis = kwargs.get('basis')
-    if basis not in EXPENSE_BASES.keys():
-        basis = 'accrualDate'
-
-    rev_basis = basis
-    if rev_basis in ['paymentDate', 'approvalDate']:
-        rev_basis = 'receivedDate'
-
-    # Used to filter the set of observations shown
-    start_date = kwargs.get('start_date')
-    end_date = kwargs.get('end_date')
-    year = kwargs.get('year')
-
-    # Needed to ensure that everything is in local time.
-    localTimeZone = None
-    if getattr(settings, 'TIME_ZONE', None):
-        localTimeZone = pytz.timezone(getattr(settings, 'TIME_ZONE'))
-
-    # Currently supported period types include month and date.  The variables
-    # values, annotations, and order_by are used repeatedly for constructing
-    # queries.
-    period_type = kwargs.get('type', 'month')
-    values = ('basisDate', )
-    order_by = ('-basisDate', )
-
-    def date_for_month(*args, **kwargs):
-        return TruncDate(TruncMonth(*args, **kwargs))
-
-    # NOTE: Django 2.1 introduces "TruncWeek", which should be added to the
-    # options once the project requires Django 2.1.
-    if period_type == 'month':
-        DateFunc = date_for_month
-    elif period_type == 'date':
-        DateFunc = TruncDate
-    else:
-        raise ValueError(_('Invalid period type passed as kwarg.'))
-
-    time_annotations = {'basisDate': DateFunc(basis, tzinfo=localTimeZone)}
-    rev_time_annotations = {'basisDate': DateFunc(rev_basis, tzinfo=localTimeZone)}
-    reg_time_annotations = {
-        'basisDate': DateFunc(
-            'eventregistration__event__startTime', tzinfo=localTimeZone
-        )
-    }
-
-    timeFilters = {'%s__isnull' % basis: False}
-    rev_timeFilters = {'%s__isnull' % rev_basis: False}
-    if start_date:
-        timeFilters['%s__gte' % basis] = start_date
-        rev_timeFilters['%s__gte' % rev_basis] = start_date
-    if end_date:
-        timeFilters['%s__lt' % basis] = end_date
-        rev_timeFilters['%s__lt' % rev_basis] = end_date
-
-    if year and not (start_date or end_date):
-        start_limit = ensure_timezone(datetime(year, 1, 1))
-        end_limit = ensure_timezone(datetime(year + 1, 1, 1))
-
-        timeFilters['%s__gte' % basis] = start_limit
-        rev_timeFilters['%s__gte' % rev_basis] = start_limit
-        timeFilters['%s__lt' % basis] = end_limit
-        rev_timeFilters['%s__lt' % rev_basis] = end_limit
-
-    # In order to provide major calculations, it's easiest to just get all the
-    # Expense and Revenue line items at once To avoid too many unnecessary
-    # calls, be sure to filter off these querysets rather than pulling them
-    # again.
-    expenseitems = ExpenseItem.objects.select_related('event').annotate(
-        **time_annotations
-    ).filter(**timeFilters)
-    revenueitems = RevenueItem.objects.select_related('event').annotate(
-        **rev_time_annotations
-    ).filter(**rev_timeFilters)
-
-    # Get the set of possible dates or months.
-    all_periods_set = set()
-
-    for qs in [expenseitems, revenueitems]:
-        all_periods_set.update(list(
-            qs.order_by().values('basisDate').annotate(n=Count('pk')).values_list('basisDate')
-        ))
-
-    all_periods = [x[0] for x in all_periods_set]
-    all_periods.sort(reverse=True)
-
-    paginator = Paginator(all_periods, kwargs.get('paginate_by', 50))
-    try:
-        paged_periods = paginator.page(kwargs.get('page', 1))
-    except PageNotAnInteger:
-        if kwargs.get('page') == 'last':
-            paged_periods = paginator.page(paginator.num_pages)
-        else:
-            paged_periods = paginator.page(1)
-    except EmptyPage:
-        paged_periods = paginator.page(paginator.num_pages)
-
-    # Define common annotations used repeatedly in queries.
-    sum_annotations = (Sum('total'), Sum('adjustments'), Sum('fees'), Sum('net'))
-
-    # Get everything by month in one query each, then pull from this.
-    totalExpensesByPeriod = expenseitems.values(*values).annotate(
-        *sum_annotations
-    ).order_by(*order_by)
-
-    instructionExpensesByPeriod = expenseitems.filter(
-        category__in=[
-            getConstant('financial__classInstructionExpenseCat'),
-            getConstant('financial__assistantClassInstructionExpenseCat')
-        ]
-    ).values(*values).annotate(*sum_annotations).order_by(*order_by)
-
-    venueExpensesByPeriod = expenseitems.filter(
-        category=getConstant('financial__venueRentalExpenseCat')
-    ).values(*values).annotate(*sum_annotations).order_by(*order_by)
-
-    totalRevenuesByPeriod = revenueitems.values(*values).annotate(
-        *sum_annotations
-    ).order_by(*order_by)
-
-    # This includes only registrations in which a series was registered for (and was not cancelled)
-    registrationsByPeriod = Registration.objects.filter(
-        eventregistration__cancelled=False
-    ).annotate(
-        **reg_time_annotations
-    ).values(*values).annotate(count=Count('id')).order_by(*order_by)
-
-    periodStatement = []
-
-    for this_period in paged_periods:
-        thisPeriodStatement = {}
-        thisPeriodStatement['period'] = this_period
-        this_period_date = datetime.combine(this_period, datetime.min.time())
-
-        if period_type == 'month':
-            thisPeriodStatement.update({
-                'period_date': this_period_date,
-                'period_name': this_period_date.strftime('%B %Y'),
-            })
-        elif period_type == 'date':
-            thisPeriodStatement.update({
-                'period_date': this_period_date,
-                'period_name': this_period.strftime('%b. %-d, %Y'),
-            })
-
-        def get_net(this_dict):
-            '''
-            Convenience function to calculate net value incorporating adjustments and fees.
-            '''
-            if not isinstance(this_dict, dict):
-                this_dict = {}
-            return this_dict.get('net__sum') or 0
-
-        thisPeriodStatement['revenues'] = get_net(
-            totalRevenuesByPeriod.filter(basisDate=this_period).first()
-        )
-        thisPeriodStatement['expenses'] = {
-            'total': get_net(totalExpensesByPeriod.filter(basisDate=this_period).first()),
-            'instruction': get_net(
-                instructionExpensesByPeriod.filter(basisDate=this_period).first()
-            ),
-            'venue': get_net(
-                venueExpensesByPeriod.filter(basisDate=this_period).first()
-            ),
-        }
-        thisPeriodStatement['expenses']['other'] = (
-            thisPeriodStatement['expenses']['total'] -
-            thisPeriodStatement['expenses']['instruction'] -
-            thisPeriodStatement['expenses']['venue']
-        )
-
-        thisPeriodStatement['registrations'] = (
-            registrationsByPeriod.filter(basisDate=this_period).first() or {}
-        ).get('count', 0)
-        thisPeriodStatement['net_profit'] = (
-            thisPeriodStatement['revenues'] - thisPeriodStatement['expenses']['total']
-        )
-        periodStatement.append(thisPeriodStatement)
-
-    periodStatement.sort(key=lambda x: x['period_date'], reverse=True)
-
-    # Return not just the statement, but also the paginator in the style of
-    # ListView's paginate_queryset()
-    return (paginator, paged_periods, periodStatement, paged_periods.has_other_pages())
-
-
-def prepareStatementByEvent(paginate=True, **kwargs):
-    all_events = Event.objects.prefetch_related(
-        'expenseitem_set', 'expenseitem_set__category',
-        'revenueitem_set', 'revenueitem_set__category',
-        'eventoccurrence_set', 'eventoccurrence_set__related_expenses__item',
-        'eventstaffmember_set', 'eventstaffmember_set__related_expenses__item',
-    )
-
-    # These are needed to allocate non-hourly expenses across events.
-    eventoccurrence_ct = ContentType.objects.get_for_model(EventOccurrence).id
-    eventstaffmember_ct = ContentType.objects.get_for_model(EventStaffMember).id
-
-    start_date = kwargs.get('start_date')
-    end_date = kwargs.get('end_date')
-    year = kwargs.get('year')
-
-    if start_date:
-        all_events = all_events.filter(year__gte=start_date.year).exclude(
-            year=start_date.year, month__lt=start_date.month
-        )
-    if end_date:
-        all_events = all_events.filter(year__lte=end_date.year).exclude(
-            year=end_date.year, month__gt=end_date.month
-        )
-    if year and not (start_date or end_date):
-        all_events = all_events.filter(year=year)
-
-
-    if paginate:
-        paginator = Paginator(all_events, kwargs.get('paginate_by', 50))
-        try:
-            paged_events = paginator.page(kwargs.get('page', 1))
-        except PageNotAnInteger:
-            if kwargs.get('page') == 'last':
-                paged_events = paginator.page(paginator.num_pages)
-            else:
-                paged_events = paginator.page(1)
-        except EmptyPage:
-            paged_events = paginator.page(paginator.num_pages)
-    else:
-        paged_events = all_events
-
-    statementByEvent = []
-
-    for event in paged_events:
-        this_event_statement = {}
-
-        # Allows access of arbitrary properties
-        this_event_statement['event'] = event
-        if event.month:
-            this_event_statement['month_name'] = '%s %s' % (month_name[event.month], event.year)
-        else:
-            this_event_statement['month_name'] = _('Unspecified Month')
-        this_event_statement['event_name'] = event.name
-        this_event_statement['registrations'] = {'total': event.numRegistered, }
-        this_event_statement['registrations'].update(event.numRegisteredByRole)
-
-        # The calculation of net vs. gross revenue for each registration item is
-        # done in models.py via model methods.  Any discounts are applied
-        # equally to each event.
-        event_revs = event.revenueitem_set.aggregate(
-            Sum('grossTotal'), Sum('total'), Sum('adjustments'), Sum('fees'), Sum('net')
-        )
-
-        this_event_statement['revenues'] = {
-            'gross': event_revs['grossTotal__sum'] or 0,
-            'netOfDiscounts': event_revs['total__sum'] or 0,
-            'adjustments': event_revs['adjustments__sum'] or 0,
-            'fees': event_revs['fees__sum'] or 0,
-            'net': event_revs['net__sum'] or 0,
-        }
-
-        allocated_venue_expenses = ExpenseItem.objects.filter(
-            event__isnull=True,
-            expensepurpose__content_type=eventoccurrence_ct,
-            expensepurpose__object_id__in=event.eventoccurrence_set.values_list('id', flat=True),
-            payTo__location__isnull=False
-        )
-
-        total_allocated_venue = sum([
-            x.getAllocationForEvents([event,]) * x.total
-            for x in allocated_venue_expenses
-        ])
-
-        allocated_staff_expenses = ExpenseItem.objects.filter(
-            event__isnull=True,
-            expensepurpose__content_type=eventstaffmember_ct,
-            expensepurpose__object_id__in=event.eventstaffmember_set.values_list('id', flat=True)
-        )
-
-        total_allocated_staff = sum([
-            x.getAllocationForEvents([event,]) * x.total
-            for x in allocated_staff_expenses
-        ])
-
-        this_event_statement['expenses'] = {
-            'instruction': event.expenseitem_set.filter(
-                category=getConstant('financial__classInstructionExpenseCat')
-            ).aggregate(Sum('total'))['total__sum'] or 0,
-            'direct_venue': event.expenseitem_set.filter(
-                category=getConstant('financial__venueRentalExpenseCat')
-            ).aggregate(Sum('total'))['total__sum'] or 0,
-            'allocated_venue': total_allocated_venue,
-            'allocated_staff': total_allocated_staff,
-            'direct_other': event.expenseitem_set.exclude(
-                category=getConstant('financial__venueRentalExpenseCat')
-            ).exclude(
-                category=getConstant('financial__classInstructionExpenseCat')
-            ).aggregate(Sum('total'))['total__sum'] or 0,
-            'fees': event.expenseitem_set.aggregate(Sum('fees'))['fees__sum'] or 0,
-        }
-        this_event_statement['expenses'].update({
-            'venue': sum([
-                this_event_statement['expenses']['direct_venue'],
-                this_event_statement['expenses']['allocated_venue'],
-            ]),
-            'other': sum([
-                this_event_statement['expenses']['direct_other'],
-                this_event_statement['expenses']['allocated_staff'],
-            ])
-        })
-        
-        this_event_statement['expenses']['total'] = sum([
-            this_event_statement['expenses']['instruction'],
-            this_event_statement['expenses']['venue'],
-            this_event_statement['expenses']['other'],
-            this_event_statement['expenses']['fees']
-        ])
-        this_event_statement['net_profit'] = (
-            this_event_statement['revenues']['net'] - this_event_statement['expenses']['total']
-        )
-
-        statementByEvent.append(this_event_statement)
-
-    if paginate:
-        # Return not just the statement, but also the paginator in the style of
-        # ListView's paginate_queryset()
-        return (paginator, paged_events, statementByEvent, paged_events.has_other_pages())
-    else:
-        return statementByEvent
-
-
-def prepareFinancialDetails(**kwargs):
-        context = kwargs.copy()
-        timeFilters = {}
-
-        # Determine the period over which the statement should be produced.
-        year = kwargs.get('year')
-        month = kwargs.get('month')
-        day = kwargs.get('day')
-        startDate = kwargs.get('startDate')
-        endDate = kwargs.get('endDate')
-        events = kwargs.get('events')
-        occurrences = kwargs.get('occurrences')
-        allocationBasis = kwargs.get('allocationBasis')
-        basis = kwargs.get('basis')
-
-        context.update({
-            'basis': basis,
-            'basis_name': EXPENSE_BASES[basis],
-            'rangeTitle': '',
-        })
-
-        if events:
-            timeFilters['event__in'] = events
-            if not occurrences:
-                context['rangeTitle'] += '; '.join([str(x.name) for x in events])
-    
-        if occurrences:
-            timeFilters['event__eventoccurrence__in'] = occurrences
-            context['rangeTitle'] += '; '.join([
-                '{}: {}'.format(x.event.name, x.timeDescription)
-                for x in occurrences
-            ])
-
-        if startDate:
-            timeFilters['%s__gte' % basis] = startDate
-            context['rangeType'] = 'Date Range'
-            context['rangeTitle'] += str(_('From %s ' % startDate.strftime('%b. %d, %Y')))
-        if endDate:
-            timeFilters['%s__lt' % basis] = endDate
-            context['rangeType'] = 'Date Range'
-            context['rangeTitle'] += str(_('To %s ' % endDate.strftime('%b. %d, %Y')))
-
-        if not startDate and not endDate:
-            start = None
-            delta = None
-
-            if day and month and year:
-                start = ensure_localtime(datetime(year, month, day))
-                delta = relativedelta(days=1)
-                context.update({
-                    'rangeType': 'Day',
-                    'rangeTitle': start.strftime('%B %d, %Y')
-                })
-            elif month and year:
-                start = ensure_localtime(datetime(year, month, 1))
-                delta = relativedelta(months=1)
-                context.update({
-                    'rangeType': 'Month',
-                    'rangeTitle': start.strftime('%B %Y')
-                })
-            elif events:
-                context['rangeType'] = 'Events'
-            elif year:
-                start = ensure_localtime(datetime(year, 1, 1))
-                delta = relativedelta(years=1)
-                context.update({
-                    'rangeType': 'Year',
-                    'rangeTitle': start.strftime('%Y')
-                })
-            else:
-                start = ensure_localtime(datetime(timezone.now().year, 1, 1))
-                delta = relativedelta(years=1)
-                context.update({
-                    'rangeType': 'YTD',
-                    'rangeTitle': _('Calendar Year To Date')
-                })
-
-            if start and delta:
-                timeFilters['%s__gte' % basis] = start
-                timeFilters['%s__lt' % basis] = start + delta
-
-        context['startDate'] = timeFilters.get('%s__gte' % basis)
-        context['endDate'] = timeFilters.get('%s__lt' % basis)
-
-        # Revenues are booked on receipt basis, not payment/approval basis
-        rev_timeFilters = timeFilters.copy()
-        rev_basis = basis
-
-        if basis in ['paymentDate', 'approvalDate']:
-            rev_basis = 'receivedDate'
-            if rev_timeFilters.get('%s__gte' % basis):
-                rev_timeFilters['receivedDate__gte'] = rev_timeFilters.get('%s__gte' % basis)
-                rev_timeFilters.pop('%s__gte' % basis, None)
-            if rev_timeFilters.get('%s__lt' % basis):
-                rev_timeFilters['receivedDate__lt'] = rev_timeFilters.get('%s__lt' % basis)
-                rev_timeFilters.pop('%s__lt' % basis, None)
-
-        expenseItems = list(
-            ExpenseItem.objects.filter(**timeFilters).annotate(
-                basisDate=Min(basis)
-            ).select_related(
-                'category', 'payTo', 'payTo__location', 'event'
-            ).prefetch_related(
-                'expensepurpose_set', 'expensepurpose_set__content_type',
-                'event__eventoccurrence_set'
-            ).order_by(basis)
-        )
-        revenueItems = list(
-            RevenueItem.objects.filter(**rev_timeFilters).annotate(
-                basisDate=Min(rev_basis)
-            ).select_related(
-                'category', 'invoiceItem', 'receivedFrom'
-            ).prefetch_related(
-                'event__eventoccurrence_set'
-            ).order_by(rev_basis)
-        )
-
-        # select_related('event') returns bare base Event instances, bypassing
-        # django-polymorphic dispatch.  Re-fetch using the polymorphic manager so
-        # that each item.event is the correct subclass (Series, PublicEvent, etc.)
-        # and its __str__ / name reflect the actual event name.
-        revenue_event_ids = list({item.event_id for item in revenueItems if item.event_id})
-        if revenue_event_ids:
-            real_events = {
-                e.id: e for e in Event.objects.filter(id__in=revenue_event_ids)
-            }
-            for item in revenueItems:
-                if item.event_id:
-                    item.event = real_events.get(item.event_id, item.event)
-
-        context['expenseItems'] = expenseItems
-        context['revenueItems'] = revenueItems
-
-        if events:
-            # These are needed to allocate non-hourly expenses across events.
-            eventoccurrence_ct = ContentType.objects.get_for_model(EventOccurrence).id
-            eventstaffmember_ct = ContentType.objects.get_for_model(EventStaffMember).id
-
-            if occurrences:
-                purpose_occurrences = occurrences.values_list('id', flat=True)
-            else:
-                purpose_occurrences = EventOccurrence.objects.filter(
-                    event__in=events
-                ).values_list('id', flat=True)
-
-            purpose_staff = EventStaffMember.objects.filter(
-                event__in=events
-            ).values_list('id', flat=True)
-            
-            context.update({
-                'allocatedVenueExpenseItems': ExpenseItem.objects.filter(
-                    event__isnull=True,
-                    expensepurpose__content_type=eventoccurrence_ct,
-                    expensepurpose__object_id__in=purpose_occurrences,
-                    payTo__location__isnull=False
-                ).annotate(
-                    basisDate=Min(basis)
-                ).order_by(basis),
-                'allocatedStaffExpenseItems': ExpenseItem.objects.filter(
-                event__isnull=True,
-                expensepurpose__content_type=eventstaffmember_ct,
-                expensepurpose__object_id__in=purpose_staff,
-                ).annotate(
-                    basisDate=Min(basis)
-                ).order_by(basis),
-            })
-            context.update({
-                'allocatedVenueTotal': sum([
-                    x.getAllocation(**allocationBasis) * (x.net or 0)
-                    for x in context['allocatedVenueExpenseItems']
-                ]),
-                'allocatedStaffTotal': sum([
-                    x.getAllocation(**allocationBasis) * (x.net or 0)
-                    for x in context['allocatedStaffExpenseItems']
-                ])
-            })
-            context['allocatedTotal'] = (
-                context['allocatedVenueTotal'] + context['allocatedStaffTotal']
-            )
-
-        # Registration revenues, instruction and venue expenses
-        # are broken out separately.
-        instruction_cats = [
-            getConstant('financial__classInstructionExpenseCat'),
-            getConstant('financial__assistantClassInstructionExpenseCat')
-        ]
-        venue_cat = getConstant('financial__venueRentalExpenseCat')
-        reg_rev_cat = getConstant('financial__registrationsRevenueCat')
-
-        context.update({
-            'instructionExpenseItems': sorted(
-                [x for x in expenseItems if x.category in instruction_cats],
-                key=lambda x: getattr(x.payTo, 'name', '')
-            ),
-            'venueExpenseItems': sorted(
-                [x for x in expenseItems if x.category == venue_cat],
-                key=lambda x: getattr(x.payTo, 'name', '')
-            ),
-            'otherExpenseItems': sorted(
-                [
-                    x for x in expenseItems if
-                    x.category not in (instruction_cats + [venue_cat])
-                ],
-                key=lambda x: getattr(x.category, 'name', '')
-            ),
-            'totalExpenses': (
-                sum([
-                    x.getAllocation(**allocationBasis) * (x.net or 0)
-                    for x in expenseItems
-                ])
-            ),
-        })
-
-        context.update({
-            'registrationRevenueItems': sorted(
-                [x for x in revenueItems if x.category == reg_rev_cat],
-                key=lambda x: (
-                    getattr(x.event, 'startTime', ensure_timezone(datetime.min)),
-                    getattr(x.event, 'uuid', '')
-                ),
-                reverse=True
-            ),
-            'otherRevenueItems': sorted(
-                [x for x in revenueItems if x.category != reg_rev_cat],
-                key=lambda x: getattr(x.category, 'name', '')
-            ),
-            'totalRevenues': sum([
-                x.getAllocation(**allocationBasis) * (x.net or 0) for x in revenueItems
-            ]),
-        })
-
-        context.update({
-            'netProfit': context['totalRevenues'] - context['totalExpenses'],
-        })
-
-        return context
