+++ danceschool/financial/management/commands/create_financial_items.py
@@ -1,8 +1,7 @@
 from django.core.management.base import BaseCommand
 
-from danceschool.financial.helpers import (
-    createExpenseItemsForEvents, createExpenseItemsForVenueRental, createRevenueItemsForRegistrations
-)
+from danceschool.financial.helpers.expenses import createExpenseItemsForEvents, createExpenseItemsForVenueRental
+from danceschool.financial.helpers.revenues import createRevenueItemsForRegistrations
 from danceschool.core.constants import getConstant
